pyproject.toml
[tool.poetry]
name = "langfuse-chat"
version = "0.1.0"
description = "Chat application with LangChain and Langfuse"
authors = ["Abe Tomoki"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
langchain = "^0.3.13"
langserve = "^0.3.0"
langfuse = "^2.57.1"
openai = "^1.58.1"
httpx = "0.24.1"
langchain-community = "^0.3.13"
langchain-openai = "^0.2.14"
sse-starlette = "^2.2.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
